requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-admin-jqueryui==1.8.22
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.15.1
django-silk==4.1.0
django-wysiwyg-redactor==0.5.1
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
Pillow==7.2.0
pycodestyle==2.6.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
